// frontend/src/app/api/models/workplace-create.ts
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

/**
 * The order for all workplaces at once via PUT /workplaces/order. A new workplace lands at
 * the end of its area.
 */
export interface WorkplaceCreate {
  areaId: string;
  blocksWorkplaceIds?: Array<string>;
  blocksWorkplacesWithTag?: Array<string>;
  description?: string;
  id: string;
  location?: string;
  maxBookingDurationMinutes?: number | null;
  name: string;
  status: 'OK' | 'DEFECT' | 'DISABLED';
  tags?: Array<string>;
  usageRules?: string | null;
  wikiUrl?: string | null;
}